Fix get_all_sessions
